refactor: 앱버전 조회 관련 코드 구조 변경 및 코드 정리
