--- src/main/java/com/recipe/app/src/appVersion/AppVersionRepositiory.java
@@ -1,8 +1,0 @@
-package com.recipe.app.src.appVersion;
-
-import com.recipe.app.src.appVersion.models.AppVersion;
-import org.springframework.data.repository.CrudRepository;
-
-public interface AppVersionRepositiory extends CrudRepository<AppVersion, Integer> {
-    AppVersion findByIdx(Integer idx);
-}
