--- src/main/java/com/recipe/app/src/appVersion/models/GetAppVersionRes.java
@@ -1,10 +1,0 @@
-package com.recipe.app.src.appVersion.models;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-
-@Getter
-@AllArgsConstructor
-public class GetAppVersionRes {
-    private final String version;
-}
