+++ src/main/java/com/recipe/app/src/common/domain/AppVersion.java
@@ -1,4 +1,4 @@
-package com.recipe.app.src.appVersion.models;
+package com.recipe.app.src.common.domain;
 
 import lombok.AccessLevel;
 import lombok.Data;
@@ -8,22 +8,21 @@ import lombok.NoArgsConstructor;
 import javax.persistence.*;
 
 
-@NoArgsConstructor(access = AccessLevel.PUBLIC) // Unit Test 를 위해 PUBLIC
+@NoArgsConstructor(access = AccessLevel.PUBLIC)
 @EqualsAndHashCode(callSuper = false)
-@Data // from lombok
-@Entity // 필수, Class 를 Database Table화 해주는 것이다
-@Table(name = "AppVersion") // Table 이름을 명시해주지 않으면 class 이름을 Table 이름으로 대체한다.
+@Data
+@Entity
+@Table(name = "AppVersion")
 public class AppVersion {
-    @Id // PK를 의미하는 어노테이션
+    @Id
     @Column(name = "idx", nullable = false, updatable = false)
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer idx;
+    private int idx;
 
     @Column(name = "version", nullable = false, length = 20)
     private String version;
 
     public AppVersion(String version) {
         this.version = version;
-
     }
 }
